Polish documentation and interfaces. (#32999)
