+++ framework/src/problems/FEProblemBase.C
@@ -4760,13 +4760,6 @@ FEProblemBase::addFVGradientMethod(const std::string & method_type,
   }
 }
 
-void
-FEProblemBase::addFVGradientMethod(const std::string & method_type, const std::string & name)
-{
-  InputParameters parameters = _factory.getValidParams(method_type);
-  addFVGradientMethod(method_type, name, parameters);
-}
-
 const UserObject &
 FEProblemBase::getUserObjectBase(const std::string & name, const THREAD_ID tid /* = 0 */) const
 {
