+++ framework/src/systems/LinearFVGradientInterface.C
@@ -24,61 +24,6 @@
 
 using namespace libMesh;
 
-LinearFVGradientReader::LinearFVGradientReader(const SystemBase & sys,
-                                               const GradientContainer & components,
-                                               const FVGradientMethod & method,
-                                               const unsigned int variable_number)
-  : _sys(sys),
-    _system_number(sys.number()),
-    _components(components),
-    _method(method),
-    _variable_number(variable_number)
-{
-}
-
-Real
-LinearFVGradientReader::component(const ElemInfo & elem_info, const unsigned int component) const
-{
-  mooseAssert(component < _components.size(), "Gradient component index out of range.");
-
-  return (*_components[component])(elem_info.dofIndices()[_system_number][_variable_number]);
-}
-
-RealVectorValue
-LinearFVGradientReader::gradient(const ElemInfo & elem_info) const
-{
-  RealVectorValue value;
-
-  for (const auto component_index : make_range(_sys.mesh().dimension()))
-    value(component_index) = component(elem_info, component_index);
-
-  return value;
-}
-
-RealVectorValue
-LinearFVGradientReader::gradient(const FaceInfo & fi) const
-{
-  const auto face_type = fi.faceType(std::make_pair(_variable_number, _system_number));
-  mooseAssert(face_type != FaceInfo::VarFaceNeighbors::NEITHER,
-              "Gradient requested on a face where the variable is defined on neither side.");
-
-  const bool var_defined_on_elem = (face_type == FaceInfo::VarFaceNeighbors::BOTH) ||
-                                   (face_type == FaceInfo::VarFaceNeighbors::ELEM);
-  const auto * const elem_one = var_defined_on_elem ? fi.elemInfo() : fi.neighborInfo();
-  const auto * const elem_two = var_defined_on_elem ? fi.neighborInfo() : fi.elemInfo();
-
-  const auto elem_one_grad = gradient(*elem_one);
-
-  if (face_type == FaceInfo::VarFaceNeighbors::BOTH)
-  {
-    mooseAssert(elem_two, "Face type indicates BOTH but neighbor information is missing.");
-    const auto elem_two_grad = gradient(*elem_two);
-    return Moose::FV::linearInterpolation(elem_one_grad, elem_two_grad, fi, var_defined_on_elem);
-  }
-  else
-    return elem_one_grad;
-}
-
 LinearFVGradientReader
 LinearFVGradientInterface::registerFVGradient(const unsigned int variable_number,
                                               const FVGradientMethod & method)
@@ -123,7 +68,7 @@ LinearFVGradientInterface::computeGradients()
 
   // Keep current values unchanged until every replacement has been computed so boundary
   // conditions consistently use gradients from the previous update.
-  // BCs may use cell gradients to compute the boundary face value, which is itself used to 
+  // BCs may use cell gradients to compute the boundary face value, which is itself used to
   // compute cell gradients
   for (auto & method_container_pair : _linear_fv_gradient_container_by_method)
     computeLinearFVGradientContainer(*method_container_pair.first, method_container_pair.second);
